requirements.txt
django
djangorestframework
python-dotenv
django-cors-headers
djangorestframework-simplejwt